Move RichSpan and RichSpanStyle from model to span package
