+++ richeditor-compose/src/commonTest/kotlin/com.mohamedrejeb.richeditor/model/RichParagraphStyleTest.kt
@@ -4,6 +4,7 @@ import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.ui.text.TextRange
 import com.mohamedrejeb.richeditor.annotation.ExperimentalRichTextApi
 import com.mohamedrejeb.richeditor.paragraph.RichParagraph
+import com.mohamedrejeb.richeditor.span.RichSpan
 import kotlin.test.Test
 import kotlin.test.assertEquals
